+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - YYYY-MM-DD
 
+### Fixed
+
+- Copy guard with reference link (#2)
+
 ### Changed
 
 - Polish documentation.